Update VaultContract.sol
The above contract modification is a new and improved version of the code with new features added. One logic error is found still. The amount is to be equally distributed amongst the Users. But in the code we are unable to take the number of users of a particular vault. userarray.length will give length of the entire users of all vaults.